Refactor test_simulate_for_sbi
