+++ tests/user_input_checks_test.py
@@ -5,6 +5,7 @@ from __future__ import annotations
 
 from typing import Callable, Tuple
 
+import numpy as np
 import pytest
 import torch
 from pyknos.mdn.mdn import MultivariateGaussianMDN
@@ -13,6 +14,7 @@ from torch.distributions import Beta, Distribution, Gamma, MultivariateNormal, U
 
 from sbi.inference import NPE_A, NPE_C, simulate_for_sbi
 from sbi.inference.posteriors.direct_posterior import DirectPosterior
+from sbi.simulators import linear_gaussian
 from sbi.simulators.linear_gaussian import diagonal_linear_gaussian
 from sbi.utils import mcmc_transform, within_support
 from sbi.utils.torchutils import BoxUniform
@@ -501,3 +503,73 @@ def test_passing_custom_density_estimator(arg):
         density_estimator = arg
     prior = MultivariateNormal(torch.zeros(2), torch.eye(2))
     _ = NPE_C(prior=prior, density_estimator=density_estimator)
+
+
+@pytest.mark.parametrize(
+    "num_simulations, simulation_batch_size, num_workers, \
+        use_process_simulator",
+    [
+        (0, None, 1, True),
+        (10, None, 1, True),
+        (100, 10, 1, True),
+        (100, None, 2, True),
+        (1000, 50, 4, True),
+        (100, 10, 2, False),
+    ],
+)
+def test_simulate_for_sbi(
+    num_simulations, simulation_batch_size, num_workers, use_process_simulator
+):
+    """Test the simulate_for_sbi function with various configurations."""
+    num_dim = 3
+    likelihood_shift = -1.0 * ones(num_dim)
+    likelihood_cov = 0.3 * eye(num_dim)
+    prior = BoxUniform(-2.0 * ones(num_dim), 2.0 * ones(num_dim))
+
+    def failing_simulator(theta):
+        if isinstance(theta, np.ndarray):
+            raise TypeError("This simulator does not support numpy arrays.")
+        return linear_gaussian(theta, likelihood_shift, likelihood_cov)
+
+    simulator = (
+        failing_simulator
+        if not use_process_simulator
+        else process_simulator(failing_simulator, prior, False)
+    )
+
+    if num_simulations == 0:
+        theta, x = simulate_for_sbi(
+            simulator=simulator,
+            proposal=prior,
+            num_simulations=num_simulations,
+            simulation_batch_size=simulation_batch_size,
+            num_workers=num_workers,
+        )
+        assert (
+            theta.numel() == 0
+        ), "Theta should be an empty tensor when num_simulations=0"
+        assert x.numel() == 0, "x should be an empty tensor when num_simulations=0"
+    else:
+        if not use_process_simulator and num_workers > 1:
+            with pytest.raises(TypeError, match="For multiprocessing"):
+                simulate_for_sbi(
+                    simulator=simulator,
+                    proposal=prior,
+                    num_simulations=num_simulations,
+                    simulation_batch_size=simulation_batch_size,
+                    num_workers=num_workers,
+                )
+        else:
+            theta, x = simulate_for_sbi(
+                simulator=simulator,
+                proposal=prior,
+                num_simulations=num_simulations,
+                simulation_batch_size=simulation_batch_size,
+                num_workers=num_workers,
+            )
+            assert (
+                theta.shape[0] == num_simulations
+            ), "Theta should have num_simulations rows"
+            assert x.shape[0] == num_simulations, "x should have num_simulations rows"
+            assert theta.shape[1] == num_dim, "Theta should have num_dim columns"
+            assert x.shape[1] == num_dim, "x should have num_dim columns"
